Move JavaTarget to JvmTarget
