+++ buildSrc/src/main/groovy/com/uber/okbuck/composer/java/JavaTestRuleComposer.groovy
@@ -3,7 +3,7 @@ package com.uber.okbuck.composer.java
 import com.google.common.collect.ImmutableSet
 import com.uber.okbuck.composer.jvm.JvmBuckRuleComposer
 import com.uber.okbuck.core.model.base.RuleType
-import com.uber.okbuck.core.model.java.JavaLibTarget
+import com.uber.okbuck.core.model.jvm.JvmTarget
 import com.uber.okbuck.template.core.Rule
 import com.uber.okbuck.template.java.JavaRule
 
@@ -15,7 +15,7 @@ final class JavaTestRuleComposer extends JvmBuckRuleComposer {
         // no instance
     }
 
-    static Rule compose(JavaLibTarget target,
+    static Rule compose(JvmTarget target,
                         RuleType ruleType = RuleType.JAVA_TEST) {
 
         List<String> deps = []
